+++ lib/src/wallet/bip32_ed25519_wallet.dart
@@ -82,7 +82,7 @@ class Bip32Ed25519Wallet implements Wallet {
   @override
   Future<List<SignatureScheme>> getSupportedSignatureSchemes(
       String keyId) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     return keyPair.supportedSignatureSchemes;
   }
 
@@ -92,7 +92,7 @@ class Bip32Ed25519Wallet implements Wallet {
     required String keyId,
     SignatureScheme? signatureScheme,
   }) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     return keyPair.sign(data, signatureScheme: signatureScheme);
   }
 
@@ -103,7 +103,7 @@ class Bip32Ed25519Wallet implements Wallet {
     required String keyId,
     SignatureScheme? signatureScheme,
   }) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     return keyPair.verify(
       data,
       signature,
@@ -149,7 +149,7 @@ class Bip32Ed25519Wallet implements Wallet {
 
   @override
   Future<PublicKey> getPublicKey(String keyId) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     final keyData = keyPair.publicKey;
     return PublicKey(keyData.id, keyData.bytes, keyData.type);
   }
@@ -164,7 +164,7 @@ class Bip32Ed25519Wallet implements Wallet {
   ///
   /// Returns a [Future] that completes with the X25519 public key as a [Uint8List].
   Future<Uint8List> getX25519PublicKey(String keyId) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     final x25519PublicKey = await keyPair.ed25519KeyToX25519PublicKey();
     return Uint8List.fromList(x25519PublicKey.bytes);
   }
@@ -175,7 +175,7 @@ class Bip32Ed25519Wallet implements Wallet {
     required String keyId,
     Uint8List? publicKey,
   }) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     return keyPair.encrypt(data, publicKey: publicKey);
   }
 
@@ -185,11 +185,12 @@ class Bip32Ed25519Wallet implements Wallet {
     required String keyId,
     Uint8List? publicKey,
   }) async {
-    final keyPair = await _getKeyPair(keyId);
+    final keyPair = await getKeyPair(keyId);
     return keyPair.decrypt(data, publicKey: publicKey);
   }
 
-  Future<Ed25519KeyPair> _getKeyPair(String keyId) async {
+  @override
+  Future<Ed25519KeyPair> getKeyPair(String keyId) async {
     if (_runtimeCache.containsKey(keyId)) {
       return _runtimeCache[keyId]!;
     }